pyproject.toml
[project]
name = "screening-api-core"
version = "0.1.0"
description = "採用スクリーニングAPI - オニオンアーキテクチャに基づくRESTful APIバックエンド"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
    "pyyaml>=6.0.0",
]
